UI, collectables and other fixes
A set of basic collectables alongside a UI detailing them has been created, as well as various changes and fixes to amend the state of the animations. As well as this, Mathf.Abs has been used to condense the check for grounded movement and a PhysicsMaterial2D has been used so that the player does not stick to a wall or side of a platform.